--- a/crates/bevy_pbr/src/render/mesh.rs
+++ b/crates/bevy_pbr/src/render/mesh.rs
@@ -25,7 +25,6 @@ use bevy_render::{
     RenderApp, RenderStage,
 };
 use bevy_transform::components::GlobalTransform;
-use smallvec::SmallVec;
 use std::num::NonZeroU64;
 
 #[derive(Default)]
@@ -191,8 +190,7 @@ impl SkinnedMeshJoints {
         let inverse_bindposes = inverse_bindposes.get(&skin.inverse_bindposes)?;
         let bindposes = inverse_bindposes.iter();
         let skin_joints = skin.joints.iter();
-        let mut temp =
-            SmallVec::<[Mat4; MAX_JOINTS]>::with_capacity(bindposes.len().min(MAX_JOINTS));
+        let mut temp = Vec::<Mat4>::with_capacity(bindposes.len().min(MAX_JOINTS));
         for (inverse_bindpose, joint) in bindposes.zip(skin_joints).take(MAX_JOINTS) {
             let joint_matrix = joints.get(*joint).ok()?.compute_matrix();
             temp.push(joint_matrix * *inverse_bindpose);