--- a/crates/bevy_pbr/src/render/mesh.rs
+++ b/crates/bevy_pbr/src/render/mesh.rs
@@ -192,8 +192,7 @@ impl SkinnedMeshJoints {
         let skin_joints = skin.joints.iter();
         let mut temp = Vec::<Mat4>::with_capacity(bindposes.len().min(MAX_JOINTS));
         for (inverse_bindpose, joint) in bindposes.zip(skin_joints).take(MAX_JOINTS) {
-            let joint_matrix = joints.get(*joint).ok()?.compute_matrix();
-            temp.push(joint_matrix * *inverse_bindpose);
+            temp.push(joints.get(*joint).ok()?.compute_matrix() * *inverse_bindpose);
         }
 
         let start = buffer.len();