--- a/crates/bevy_pbr/src/render/mesh.rs
+++ b/crates/bevy_pbr/src/render/mesh.rs
@@ -190,13 +190,16 @@ impl SkinnedMeshJoints {
         let inverse_bindposes = inverse_bindposes.get(&skin.inverse_bindposes)?;
         let bindposes = inverse_bindposes.iter();
         let skin_joints = skin.joints.iter();
-        let mut temp = Vec::<Mat4>::with_capacity(bindposes.len().min(MAX_JOINTS));
+        let start = buffer.len();
         for (inverse_bindpose, joint) in bindposes.zip(skin_joints).take(MAX_JOINTS) {
-            temp.push(joints.get(*joint).ok()?.compute_affine() * *inverse_bindpose);
+            if let Ok(joint) = joints.get(*joint) {
+                buffer.push(joint.compute_affine() * *inverse_bindpose);
+            } else {
+                buffer.truncate(start);
+                return None;
+            }
         }
 
-        let start = buffer.len();
-        buffer.extend(temp);
         // Pad to 256 byte alignment
         while buffer.len() % 4 != 0 {
             buffer.push(Mat4::ZERO);