@superdump 2022-04-05
I’d be interested to make. GlobalTransform just an Affine3A internally and not a separate representation at all, and see what impact that has on relevant systems… :)