@mockersf 2022-04-05
yup that would probably be faster, but also a much bigger bite to take...