@superdump 2022-04-05
Also, the title is incorrect. An almost 48% reduction in execution time is enormous.