@james7132 2022-04-05
Copying my comments from Discord:

>[2:12 PM] james7132: the temp buffer is only added in the case of a failure to fetch.
>[2:12 PM] james7132: if we can either remove the failure case
>[2:12 PM] james7132: or find another way to fail without breaking the state of the total joint buffer, we could probably save quite a bit of time

if we can avoid this double buffered copy, I think we could save quite a bit more time too.